+++ src/components/refactor/dialogs/two-windings-transformer-creation/tap-changer-pane/create-rule/create-rule-dialog-utils.js
