+++ src/components/refactor/dialogs/two-windings-transformer-creation/tap-changer-pane/create-rule/create-rule-form.js
@@ -9,7 +9,7 @@ import { FormattedMessage } from 'react-intl';
 import {
     PHASE_TAP,
     RATIO_TAP,
-} from '../two-windings-transformer-creation-dialog';
+} from '../../two-windings-transformer-creation-dialog';
 
 const CreateRuleForm = (props) => {
     const computeLowTapLabel = (ruleType) => {
