+++ src/components/refactor/dialogs/two-windings-transformer-creation/tap-changer-pane/phase-tap-changer-pane/phase-tap-changer-pane-taps.js
@@ -17,8 +17,8 @@ import {
 } from 'components/refactor/utils/field-constants';
 import React, { useMemo } from 'react';
 import { useIntl } from 'react-intl';
-import TapChangerPaneTaps from '../tap-changer-pane-table/tap-changer-pane-taps';
-import { PHASE_TAP } from '../two-windings-transformer-creation-dialog';
+import TapChangerPaneTaps from '../tap-changer-pane-taps';
+import { PHASE_TAP } from '../../two-windings-transformer-creation-dialog';
 
 const PhaseTapChangerPaneTaps = ({ disabled }) => {
     const intl = useIntl();
